+++ lib/rxflare.dart
@@ -3,6 +3,7 @@ export 'src/rx.dart';
 export 'src/rx_debug.dart';
 export 'src/rx_event_bus.dart';
 export 'src/rx_obj_mgr.dart';
+export 'src/rx_builder.dart';
 
 export 'src/rx_state.dart';
 export 'src/rx_stack.dart';